Refactored (2)
Created an `Editor` class out methods in the `Util` class to enhance
code clarity.
