+++ cheat/sheet.py
@@ -2,14 +2,16 @@ import io
 import os
 import shutil
 
+from cheat.editor import Editor
 from cheat.utils import Utils
 
 
 class Sheet:
 
-    def __init__(self, sheets, utils):
+    def __init__(self, sheets, utils, editor):
         self._sheets = sheets
         self._utils = utils
+        self._editor = editor
 
     def copy(self, current_sheet_path, new_sheet_path):
         """ Copies a sheet to a new path """
@@ -44,11 +46,11 @@ class Sheet:
     def create(self, sheet):
         """ Creates a cheatsheet """
         new_sheet_path = os.path.join(self._sheets.default_path(), sheet)
-        self._utils.open_with_editor(new_sheet_path)
+        self._editor.open(new_sheet_path)
 
     def edit(self, sheet):
         """ Opens a cheatsheet for editing """
-        self._utils.open_with_editor(self.path(sheet))
+        self._editor.open(self.path(sheet))
 
     def exists(self, sheet):
         """ Predicate that returns true if the sheet exists """
