+++ cheat/utils.py
@@ -68,26 +68,6 @@ class Utils:
         Utils.warn(message)
         exit(1)
 
-    def editor(self):
-        """ Determines the user's preferred editor """
-
-        # assert that the editor is set
-        if (not self._config.cheat_editor):
-            Utils.die(
-                'You must set a CHEAT_EDITOR, VISUAL, or EDITOR environment '
-                'variable or setting in order to create/edit a cheatsheet.'
-            )
-
-        return self._config.cheat_editor
-
-    def open_with_editor(self, filepath):
-        """ Open `filepath` using the EDITOR specified by the env variables """
-        editor_cmd = self.editor().split()
-        try:
-            subprocess.call(editor_cmd + [filepath])
-        except OSError:
-            Utils.die('Could not launch ' + self.editor())
-
     @staticmethod
     def warn(message):
         """ Prints a message to stderr """
